organized the order.js file
